requirements.txt
asgiref==3.5.2
beautifulsoup4==4.8.2
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
crypto==1.4.1
Django==4.0.4
django-jinja==2.10.2
Flask==2.1.2
Flask-MySQLdb==1.0.1
idna==2.8
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
key-generator==1.0.3
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.0
Naked==0.1.31
nltk==3.4.5
pyasn1==0.4.8
pycurl==7.45.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.22.0
rsa==4.8
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tk==0.1.0
urllib3==1.25.11
Werkzeug==2.1.2
zipp==3.8.0

driver~=0.1.0
robotframework-selenium2library~=3.0.0
selenium~=4.1.5
furl~=2.1.3
APScheduler~=3.9.1